requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
cloudinary==1.30.0
contourpy==1.0.6
cycler==0.11.0
Django==4.1.4
et-xmlfile==1.1.0
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
kiwisolver==1.4.4
line-bot-sdk==2.3.0
matplotlib==3.6.2
multidict==6.0.3
numpy==1.24.0
openai==0.25.0
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
pandas-stubs==1.5.2.221213
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tqdm==4.64.1
types-pytz==2022.7.0.0
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
